+++ src/main/java/uit/team/QLSachJFrame1.java
@@ -52,6 +52,7 @@ public class QLSachJFrame1 extends QLFrame {
         modify = new javax.swing.JButton();
         delete = new javax.swing.JButton();
         jLabel1 = new javax.swing.JLabel();
+        errorLabel1 = new javax.swing.JLabel();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -93,6 +94,9 @@ public class QLSachJFrame1 extends QLFrame {
         jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
         jLabel1.setText("Quản lý sách");
 
+        errorLabel1.setFont(new java.awt.Font("Dialog", 3, 14)); // NOI18N
+        errorLabel1.setForeground(new java.awt.Color(255, 0, 0));
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -102,10 +106,12 @@ public class QLSachJFrame1 extends QLFrame {
                 .addComponent(insertAction, javax.swing.GroupLayout.PREFERRED_SIZE, 116, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(220, 220, 220)
                 .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 402, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap(305, Short.MAX_VALUE))
             .addComponent(jScrollPane1)
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addContainerGap(659, Short.MAX_VALUE)
+                .addGap(24, 24, 24)
+                .addComponent(errorLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 540, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
                 .addComponent(modify)
                 .addGap(165, 165, 165)
                 .addComponent(delete)
@@ -120,11 +126,15 @@ public class QLSachJFrame1 extends QLFrame {
                     .addComponent(insertAction, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE))
                 .addGap(36, 36, 36)
                 .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 537, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(modify)
-                    .addComponent(delete))
-                .addContainerGap(17, Short.MAX_VALUE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(modify)
+                            .addComponent(delete))
+                        .addGap(0, 17, Short.MAX_VALUE))
+                    .addComponent(errorLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
         );
 
         pack();
@@ -140,18 +150,28 @@ public class QLSachJFrame1 extends QLFrame {
 
     private void modifyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_modifyActionPerformed
         // TODO add your handling code here:
-        FormInsertUpdate form = new FormSach(true);
-        form.setVisible(true);
-        this.returnHomeFrame = false;
-        this.dispose();
+        String selectedPrV = QLController.getSelectedRowFirstCell(jTable1);
+        if (selectedPrV.startsWith("ERR:")){
+            this.errorLabel1.setText(selectedPrV);
+        } else {
+            FormInsertUpdate form = new FormSach(true);
+            form.setVisible(true);
+            this.returnHomeFrame = false;
+            this.dispose();
+        }
     }//GEN-LAST:event_modifyActionPerformed
 
     private void deleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteActionPerformed
         // TODO add your handling code here:
-        JFrame deleteForm = new DeleteForm(this);
-        deleteForm.setVisible(true);
-        this.returnHomeFrame = false;
-        this.dispose();
+        String selectedPrV = QLController.getSelectedRowFirstCell(jTable1);
+        if (selectedPrV.startsWith("ERR:")){
+            this.errorLabel1.setText(selectedPrV);
+        } else {
+            JFrame deleteForm = new DeleteForm(this);
+            deleteForm.setVisible(true);
+            this.returnHomeFrame = false;
+            this.dispose();
+        }
     }//GEN-LAST:event_deleteActionPerformed
 
     /**
@@ -192,6 +212,7 @@ public class QLSachJFrame1 extends QLFrame {
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton delete;
+    private javax.swing.JLabel errorLabel1;
     private javax.swing.JButton insertAction;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JScrollPane jScrollPane1;
